+++ Changelog.md
@@ -34,6 +34,7 @@
 - Fix bug where graders can't see the tests that they run (#5210)
 - Fix bug where graders can't release results on the results page (#5210)
 - Use DOMpurify library to sanitize rendered markdown content (#5211)
+- Add percentage extra marks when calculating total extra marks properly (#5213)
 
 ## [v1.11.5]
 - Account for percentage deductions when calculating total marks after deleting a criterion (#5176)